pyproject.toml
[project]
name = "bothackathon"
version = "0.1.0"
requires-python = "~=3.13.0"
dependencies = [
    "emoji>=2.14.1",
    "google-genai>=1.3.0",
    "pymongo>=4.11.2",
    "pytelegrambotapi>=4.26.0",
]

[dependency-groups]
dev = [
    "ipython>=9.0.0",
]
